Move logger to logger.go
